refactor: add missing polars-ops tests to CI (#11859)

+++ crates/polars-ops/src/chunked_array/interpolate.rs
@@ -242,26 +242,34 @@ mod test {
     fn test_interpolate() {
         let ca = UInt32Chunked::new("", &[Some(1), None, None, Some(4), Some(5)]);
         let out = interpolate(&ca.into_series(), InterpolationMethod::Linear);
-        let out = out.u32().unwrap();
+        let out = out.f64().unwrap();
         assert_eq!(
             Vec::from(out),
-            &[Some(1), Some(2), Some(3), Some(4), Some(5)]
+            &[Some(1.0), Some(2.0), Some(3.0), Some(4.0), Some(5.0)]
         );
 
         let ca = UInt32Chunked::new("", &[None, Some(1), None, None, Some(4), Some(5)]);
         let out = interpolate(&ca.into_series(), InterpolationMethod::Linear);
-        let out = out.u32().unwrap();
+        let out = out.f64().unwrap();
         assert_eq!(
             Vec::from(out),
-            &[None, Some(1), Some(2), Some(3), Some(4), Some(5)]
+            &[None, Some(1.0), Some(2.0), Some(3.0), Some(4.0), Some(5.0)]
         );
 
         let ca = UInt32Chunked::new("", &[None, Some(1), None, None, Some(4), Some(5), None]);
         let out = interpolate(&ca.into_series(), InterpolationMethod::Linear);
-        let out = out.u32().unwrap();
+        let out = out.f64().unwrap();
         assert_eq!(
             Vec::from(out),
-            &[None, Some(1), Some(2), Some(3), Some(4), Some(5), None]
+            &[
+                None,
+                Some(1.0),
+                Some(2.0),
+                Some(3.0),
+                Some(4.0),
+                Some(5.0),
+                None
+            ]
         );
         let ca = UInt32Chunked::new("", &[None, Some(1), None, None, Some(4), Some(5), None]);
         let out = interpolate(&ca.into_series(), InterpolationMethod::Nearest);
@@ -276,8 +284,11 @@ mod test {
     fn test_interpolate_decreasing_unsigned() {
         let ca = UInt32Chunked::new("", &[Some(4), None, None, Some(1)]);
         let out = interpolate(&ca.into_series(), InterpolationMethod::Linear);
-        let out = out.u32().unwrap();
-        assert_eq!(Vec::from(out), &[Some(4), Some(3), Some(2), Some(1)])
+        let out = out.f64().unwrap();
+        assert_eq!(
+            Vec::from(out),
+            &[Some(4.0), Some(3.0), Some(2.0), Some(1.0)]
+        )
     }
 
     #[test]
